requirements.txt
arabic-reshaper==3.0.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
deep-translator==1.11.4
greenlet==3.2.3
idna==3.10
PyAudio==0.2.14
python-bidi==0.6.6
requests==2.32.4
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.5.0
